+++ frontend/controllers/PengurusanInsuranController.php
@@ -5,12 +5,24 @@ namespace frontend\controllers;
 use Yii;
 use app\models\PengurusanInsuran;
 use frontend\models\PengurusanInsuranSearch;
+use app\models\MsnLaporanTuntutanInsurans;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\UploadedFile;
+use yii\helpers\Json;
+use yii\helpers\BaseUrl;
+
+use app\models\general\Upload;
+use app\models\general\GeneralVariable;
+use common\models\general\GeneralFunction;
 
 // table reference
 use app\models\Atlet;
+use app\models\RefProgramSemasaSukanAtlet;
+use app\models\RefSukan;
+use app\models\RefJenisTuntutan;
+use app\models\RefStatusPermohonanInsuran;
 
 /**
  * PengurusanInsuranController implements the CRUD actions for PengurusanInsuran model.
@@ -35,6 +47,10 @@ class PengurusanInsuranController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
         $searchModel = new PengurusanInsuranSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
@@ -51,10 +67,26 @@ class PengurusanInsuranController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
         $model = $this->findModel($id);
         
         $ref = Atlet::findOne(['atlet_id' => $model->atlet_id]);
         $model->atlet_id = $ref['nameAndIC'];
+        
+        $ref = RefProgramSemasaSukanAtlet::findOne(['id' => $model->program]);
+        $model->program = $ref['desc'];
+        
+        $ref = RefSukan::findOne(['id' => $model->sukan]);
+        $model->sukan = $ref['desc'];
+        
+        $ref = RefJenisTuntutan::findOne(['id' => $model->jenis_tuntutan]);
+        $model->jenis_tuntutan = $ref['desc'];
+        
+        $ref = RefStatusPermohonanInsuran::findOne(['id' => $model->status_permohonan]);
+        $model->status_permohonan = $ref['desc'];
         
         return $this->render('view', [
             'model' => $model,
@@ -69,16 +101,29 @@ class PengurusanInsuranController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
         $model = new PengurusanInsuran();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->pengurusan_insuran_id]);
-        } else {
-            return $this->render('create', [
+            $file = UploadedFile::getInstance($model, 'lampiran');
+            if($file){
+                $model->lampiran = Upload::uploadFile($file, Upload::pengurusanInsuranFolder, $model->pengurusan_insuran_id);
+            }
+            
+            $model->tarikh_permohonan = $model->created; // auto capture timestamp
+            
+            if($model->save()){
+                return $this->redirect(['view', 'id' => $model->pengurusan_insuran_id]);
+            }
+        } 
+        
+        return $this->render('create', [
                 'model' => $model,
                 'readonly' => false,
             ]);
-        }
     }
 
     /**
@@ -89,10 +134,21 @@ class PengurusanInsuranController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->pengurusan_insuran_id]);
+            $file = UploadedFile::getInstance($model, 'lampiran');
+            if($file){
+                $model->lampiran = Upload::uploadFile($file, Upload::pengurusanInsuranFolder, $model->pengurusan_insuran_id);
+            }
+            
+            if($model->save()){
+                return $this->redirect(['view', 'id' => $model->pengurusan_insuran_id]);
+            }
         } else {
             return $this->render('update', [
                 'model' => $model,
@@ -109,6 +165,13 @@ class PengurusanInsuranController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
+        // delete upload file
+        self::actionDeleteupload($id, 'lampiran');
+        
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -128,5 +191,76 @@ class PengurusanInsuranController extends Controller
         } else {
             throw new NotFoundHttpException('The requested page does not exist.');
         }
+    }
+    
+    // Add function for delete image or file
+    public function actionDeleteupload($id, $field)
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
+            $img = $this->findModel($id)->$field;
+            
+            if($img){
+                if (!unlink($img)) {
+                    return false;
+                }
+            }
+
+            $img = $this->findModel($id);
+            $img->$field = NULL;
+            $img->update();
+
+            return $this->redirect(['update', 'id' => $id]);
+    }
+    
+    public function actionLaporanTuntutanInsurans()
+    {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(array(GeneralVariable::loginPagePath));
+        }
+        
+        $model = new MsnLaporanTuntutanInsurans();
+        $model->format = 'html';
+
+        if ($model->load(Yii::$app->request->post())) {
+            
+            if($model->format == "html") {
+                $report_url = BaseUrl::to(['generate-laporan-tuntutan-insurans'
+                    , 'tarikh_hingga' => $model->tarikh_hingga
+                    , 'tarikh_dari' => $model->tarikh_dari
+                    , 'format' => $model->format
+                ], true);
+                echo "<script type=\"text/javascript\" language=\"Javascript\">window.open('".$report_url."');</script>";
+            } else {
+                return $this->redirect(['generate-laporan-tuntutan-insurans'
+                    , 'tarikh_dari' => $model->tarikh_dari
+                    , 'tarikh_hingga' => $model->tarikh_hingga
+                    , 'format' => $model->format
+                ]);
+            }
+        } 
+
+        return $this->render('laporan_tuntutan_insurans', [
+            'model' => $model,
+            'readonly' => false,
+        ]);
+    }
+    
+    public function actionGenerateLaporanTuntutanInsurans($tarikh_dari, $tarikh_hingga, $format)
+    {
+        if($tarikh_dari == "") $tarikh_dari = array();
+        else $tarikh_dari = array($tarikh_dari);
+        
+        if($tarikh_hingga == "") $tarikh_hingga = array();
+        else $tarikh_hingga = array($tarikh_hingga);
+        
+        $controls = array(
+            'FROM_DATE' => $tarikh_dari,
+            'TO_DATE' => $tarikh_hingga,
+        );
+        
+        GeneralFunction::generateReport('/spsb/MSN/LaporanTuntutanInsurans', $format, $controls, 'laporan_tuntutan_insurans');
     }
 }
